Fix the camera uniform
The orbit-camera.ts was returning a Float32Array, which was being treated as a normal array later (see the case to `as number[]`)`. Unfortunately this was unsound: spreading & concating a Float32Array treats it as a single value, not an array of values. So later in uniforms.ts, `[location].concat(this.values)` (where this.values was the Float32Array), produced `[location, Float32Array]` instead of `[location, x, y, z]`, which is what the webgl api requires.

I didn't see any other usages of `toFloat32Array`, so I figured changing it to be `toArray` made sense.
I also added demos to the docs page for the camera and the mouse, to make it easy to verify that they're working as expected.